1) Make the 'Separator' class more general by moving it to widgets2 and making it usable in .ui files. 2) Change the template tester to use FlowLayout.
