+++ src/calibre/gui2/preferences/coloring.py
@@ -22,7 +22,7 @@ from calibre.gui2 import (error_dialog, choose_files, pixmap_to_data, gprefs,
                           choose_save_file, open_local_file)
 from calibre.gui2.dialogs.template_dialog import TemplateDialog
 from calibre.gui2.metadata.single_download import RichTextDelegate
-from calibre.gui2.widgets2 import ColorButton, FlowLayout
+from calibre.gui2.widgets2 import ColorButton, FlowLayout, Separator
 from calibre.library.coloring import (Rule, conditionable_columns,
     displayable_columns, rule_from_template, color_row_key)
 from calibre.utils.localization import lang_map
@@ -893,33 +893,6 @@ class RulesView(QListView):  # {{{
             _, _, rule = self.model().data(new, Qt.ItemDataRole.UserRole)
             self.enable_convert_buttons_function(isinstance(rule, Rule))
         return super().currentChanged(new, prev)
-# }}}
-
-
-class Separator(QWidget):  # {{{
-
-    def __init__(self, parent, widget_for_height):
-        QWidget.__init__(self, parent)
-        self.bcol = QColor(QPalette.ColorRole.Text)
-        self.update_brush()
-        self.widget_for_height = widget_for_height
-        self.setSizePolicy(QSizePolicy.Policy.Fixed, QSizePolicy.Policy.MinimumExpanding)
-
-    def update_brush(self):
-        self.brush = QBrush(self.bcol)
-        self.update()
-
-    def sizeHint(self):
-        return QSize(1, self.widget_for_height.height())
-
-    def paintEvent(self, ev):
-        painter = QPainter(self)
-        # Purely subjective: shorten the line a bit to look 'better'
-        r = ev.rect()
-        r.setTop(r.top() + 3)
-        r.setBottom(r.bottom() - 3)
-        painter.fillRect(r, self.brush)
-        painter.end()
 # }}}
